refactor: split Utils.kt into CredentialMapping.kt and PublicationMapping.kt
