+++ src/main/kotlin/io/github/zenhelix/gradle/plugin/utils/CredentialMapping.kt
@@ -6,17 +6,8 @@ import io.github.zenhelix.gradle.plugin.client.model.Outcome
 import io.github.zenhelix.gradle.plugin.client.model.Success
 import io.github.zenhelix.gradle.plugin.client.model.ValidationError
 import io.github.zenhelix.gradle.plugin.extension.MavenCentralUploaderExtension
-import io.github.zenhelix.gradle.plugin.extension.PublishingMode
-import io.github.zenhelix.gradle.plugin.task.ArtifactFileInfo
-import io.github.zenhelix.gradle.plugin.task.ArtifactInfo
-import io.github.zenhelix.gradle.plugin.task.CreateChecksumTask
-import io.github.zenhelix.gradle.plugin.task.GAV
-import io.github.zenhelix.gradle.plugin.task.PublicationInfo
 import org.gradle.api.Project
 import org.gradle.api.provider.Provider
-import org.gradle.api.publish.maven.internal.publication.MavenPublicationInternal
-import org.gradle.api.tasks.TaskProvider
-import org.gradle.kotlin.dsl.listProperty
 
 internal fun Project.mapCredentials(
     extension: MavenCentralUploaderExtension
@@ -50,26 +41,4 @@ internal fun Project.mapCredentials(
             Failure(ValidationError.NoCredentials)
         }
     }
-}
-
-internal fun MavenPublicationInternal.mapModel(
-    project: Project,
-    checksumTask: TaskProvider<CreateChecksumTask>
-): PublicationInfo = PublicationInfo(
-    projectPath = project.path,
-    gav = GAV.of(this),
-    publicationName = this.name,
-    artifacts = project.objects.listProperty<ArtifactInfo>().apply {
-        convention(project.provider {
-            this@mapModel.publishableArtifacts.map {
-                ArtifactInfo(artifact = ArtifactFileInfo.of(it), gav = GAV.of(this@mapModel))
-            }
-        })
-    },
-    checksumFiles = checksumTask.flatMap { it.checksumFiles }
-)
-
-internal fun PublishingMode.mapModel(): io.github.zenhelix.gradle.plugin.client.model.PublishingType = when (this) {
-    PublishingMode.AUTOMATIC -> io.github.zenhelix.gradle.plugin.client.model.PublishingType.AUTOMATIC
-    PublishingMode.USER_MANAGED -> io.github.zenhelix.gradle.plugin.client.model.PublishingType.USER_MANAGED
 }
