+++ src/main/kotlin/io/github/zenhelix/gradle/plugin/utils/PublicationMapping.kt
@@ -1,0 +1,34 @@
+package io.github.zenhelix.gradle.plugin.utils
+
+import io.github.zenhelix.gradle.plugin.extension.PublishingMode
+import io.github.zenhelix.gradle.plugin.task.ArtifactFileInfo
+import io.github.zenhelix.gradle.plugin.task.ArtifactInfo
+import io.github.zenhelix.gradle.plugin.task.CreateChecksumTask
+import io.github.zenhelix.gradle.plugin.task.GAV
+import io.github.zenhelix.gradle.plugin.task.PublicationInfo
+import org.gradle.api.Project
+import org.gradle.api.publish.maven.internal.publication.MavenPublicationInternal
+import org.gradle.api.tasks.TaskProvider
+import org.gradle.kotlin.dsl.listProperty
+
+internal fun MavenPublicationInternal.mapModel(
+    project: Project,
+    checksumTask: TaskProvider<CreateChecksumTask>
+): PublicationInfo = PublicationInfo(
+    projectPath = project.path,
+    gav = GAV.of(this),
+    publicationName = this.name,
+    artifacts = project.objects.listProperty<ArtifactInfo>().apply {
+        convention(project.provider {
+            this@mapModel.publishableArtifacts.map {
+                ArtifactInfo(artifact = ArtifactFileInfo.of(it), gav = GAV.of(this@mapModel))
+            }
+        })
+    },
+    checksumFiles = checksumTask.flatMap { it.checksumFiles }
+)
+
+internal fun PublishingMode.mapModel(): io.github.zenhelix.gradle.plugin.client.model.PublishingType = when (this) {
+    PublishingMode.AUTOMATIC -> io.github.zenhelix.gradle.plugin.client.model.PublishingType.AUTOMATIC
+    PublishingMode.USER_MANAGED -> io.github.zenhelix.gradle.plugin.client.model.PublishingType.USER_MANAGED
+}
